feat(module-federation): Update to the latest version of @module-federation/enhanced and @module-federation/node to fix koa security vulnerability
Summary

Koa uses an evil regex to parse the X-Forwarded-Proto and X-Forwarded-Host HTTP headers. This can be exploited to carry out a Denial-of-Service attack.
PoC

Coming soon.
Impact

This is a Regex Denial-of-Service attack and causes memory exhaustion. The regex should be improved and empty values should not be allowed.